UNDO delete car
* delete car only after snackbar is dismissed
