+++ app/src/main/java/cz/muni/fi/a2p06/stolencardatabase/fragments/CarDetailFragment.java
@@ -6,7 +6,6 @@ import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -25,12 +24,6 @@ import com.google.android.gms.maps.model.LatLng;
 import com.google.android.gms.maps.model.MarkerOptions;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
-import com.google.firebase.database.DataSnapshot;
-import com.google.firebase.database.DatabaseError;
-import com.google.firebase.database.DatabaseReference;
-import com.google.firebase.database.FirebaseDatabase;
-import com.google.firebase.database.Query;
-import com.google.firebase.database.ValueEventListener;
 import com.google.firebase.storage.FirebaseStorage;
 import com.google.firebase.storage.StorageReference;
 import com.stepstone.stepper.Step;
@@ -198,9 +191,8 @@ public class CarDetailFragment extends Fragment implements Step, OnMapReadyCallb
 
     @OnClick(R.id.delete_car)
     public void onDeleteCarClick(View view) {
-        deleteCar();
         if (mListener != null) {
-            mListener.onDeleteCar();
+            mListener.onDeleteCar(mCar);
         }
     }
 
@@ -209,26 +201,6 @@ public class CarDetailFragment extends Fragment implements Step, OnMapReadyCallb
         if (mListener != null) {
             mListener.onEditCar(mCar);
         }
-    }
-
-    private void deleteCar() {
-        DatabaseReference ref = FirebaseDatabase.getInstance().getReference("cars");
-        Query query = ref.orderByChild("regno").equalTo(mCar.getRegno());
-        query.addListenerForSingleValueEvent(new ValueEventListener() {
-            @Override
-            public void onDataChange(DataSnapshot dataSnapshot) {
-                for (DataSnapshot snapshot : dataSnapshot.getChildren()) {
-                    Car car = snapshot.getValue(Car.class);
-                    HelperMethods.removePhotoFromDb(car.getPhotoUrl());
-                    snapshot.getRef().removeValue();
-                }
-            }
-
-            @Override
-            public void onCancelled(DatabaseError databaseError) {
-                Log.e(TAG, "onCancelled", databaseError.toException());
-            }
-        });
     }
 
     public void updateCarView(Car car) {
@@ -337,7 +309,7 @@ public class CarDetailFragment extends Fragment implements Step, OnMapReadyCallb
     }
 
     public interface OnCarDetailFragmentInteractionListener {
-        void onDeleteCar();
+        void onDeleteCar(Car car);
         void onEditCar(Car car);
     }
 }
