+++ app/src/main/res/layout/activity_main.xml
@@ -5,6 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:id="@+id/main_activity_root_view"
     tools:context="cz.muni.fi.a2p06.stolencardatabase.MainActivity">
 
     <android.support.design.widget.AppBarLayout
